+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 ### Added
 - `PageUp` and `PageDown` keys can now be used to scroll in the MessageList
 - Keeping `Alt + ArrowUp/ArrowDown` pressed now keeps selecting the next chat until the key is released
+- Download on Demand
+
+## Changed
+- update `filesize` dependency to version `8.0.6`
 
 ## [1.25.1] - 2021-11-30